[rename] 라우트 일부 변경
