+++ components/exchange/chart/market-list/MarketListUI.jsx
@@ -20,7 +20,7 @@ const TableCell = ({
 }) => {
   return (
     <td
-      className={`table-cell ${width} px-0.5 border-b border-b-gray-100 ${align} ${changeRate < 0 ? 'text-negative' : changeRate > 0 ? 'text-positive' : 'text-black'} ${className}`}
+      className={`table-cell ${width} py-1 px-0.5 border-b border-b-gray-100 ${align} ${changeRate < 0 ? 'text-negative' : changeRate > 0 ? 'text-positive' : 'text-black'} ${className}`}
     >
       {children}
     </td>
@@ -51,10 +51,10 @@ export default function MarketList({
     <div className="flex flex-col h-full">
       <div className="relative w-full">
         <input
-          aria-label="마켓 검색"
+          aria-label="코인명 검색"
           type="text"
           className="w-full h-[2.5rem] px-1 bg-white border border-none font-ng placeholder:font-ng focus:outline-none focus:ring-1 focus:ring-main-dark"
-          placeholder="마켓 검색"
+          placeholder="코인명/심볼 검색"
           value={keyword}
           onChange={handleSearchChange}
         />
