+++ pages/mypage/history.jsx
@@ -1,0 +1,3 @@
+export default function History() {
+  return <div></div>;
+}
